Adds basic structure to AWSLambdaCost class
This class is responsible by calculate the lambda function costs

+++ AWSLambdaLab.PricingCalc/Models/ComputationParameters.cs
@@ -5,6 +5,6 @@ namespace AWSLambdaLab.PricingCalc.Models
     {
         public decimal AllocatedMemoryByFunctionInMegaBytes { get; set; }
         public int NumberOfRequests { get; set; }
-        public int TotalComputeByFunctionInSeconds { get; set; }
+        public decimal TotalComputeByFunctionInSeconds { get; set; }
     }
 }
